added Worker Items Services on the Worker Feature

+++ OilProduction.Application/Features/WorkService/Commands/CreateWork/CreateWorkCommand.cs
@@ -3,7 +3,7 @@
 
 namespace OilProduction.Application.Features.WorkService.Commands.CreateWork
 {
-    public class CreateWorkCommand : IRequest<int>
+    public class CreateWorkCommand : IRequest<long>
     {
         public string Name { get; set; }
         public string Address { get; set; }
